replace the setoutput with new version

+++ entrypoint.sh
@@ -3,8 +3,14 @@
 set -e
 
 echo -e "\033[36mSetting up kubectl configuration\033[0m"
-mkdir -p ~/.kube/
-echo "${INPUT_KUBECONFIG}" > ~/.kube/config
+
+mkdir ~/.kube/ || true
+if [ -f "~/.kube/config" ]; then
+  echo -e "\033[36mExisting kubeconfig found, using that and ignoring input\033[0m"
+else
+  echo -e "\033[36mUsing kubeconfig from input\033[0m"
+  echo "${INPUT_KUBECONFIG}" > ~/.kube/config
+fi
 
 echo -e "\033[36mPreparing helm execution\033[0m"
 echo "${INPUT_EXEC}" > run.sh
@@ -14,11 +20,10 @@ echo -e "\033[36mExecuting helm\033[0m"
 helm_output=$(./run.sh)
 echo "$helm_output"
 
-helm_output="${helm_output//'%'/'%25'}"
-helm_output="${helm_output//$'\n'/'%0A'}"
-helm_output="${helm_output//$'\r'/'%0D'}"
-
-echo "helm_output=${helm_output}" >> "${GITHUB_OUTPUT}"
+delimiter="$(openssl rand -hex 8)"
+echo "helm_output<<${delimiter}" >> "${GITHUB_OUTPUT}"
+echo "$helm_output" >> "${GITHUB_OUTPUT}"
+echo "${delimiter}" >> "${GITHUB_OUTPUT}"
 
 echo -e "\033[36mCleaning up: \033[0m"
 rm ./run.sh -Rf
